p4i integration Integrate 180991 from 4.0 to main: Fixes:    Make the font engine's drawText() call take qreals. RevBy:    Gunnar Details:  If we don't want bad truncation when we are drawing text, these functions should take qreals.
[git-p4: depot-paths = "//depot/qt/main/": change = 180992]

+++ src/gui/text/qfontengine.cpp
@@ -119,7 +119,7 @@ glyph_metrics_t QFontEngineBox::boundingBox(const QGlyphLayout *, int numGlyphs)
 }
 
 #if !defined(Q_WS_X11) && !defined(Q_WS_WIN)
-void QFontEngineBox::draw(QPaintEngine *p, int x, int y, const QTextItemInt &si)
+void QFontEngineBox::draw(QPaintEngine *p, qreal x, qreal y, const QTextItemInt &si)
 {
     Q_UNUSED(p);
     Q_UNUSED(x);
